Refactored item endpoints to now get positional information at point of initialization, moved all game objects to a game object class, game objects now inherit from same parent
