+++ text_adventure/text_adventure_game.py
@@ -1,289 +1,9 @@
 from game import Game
-from dataclasses import dataclass
 import json
 import random
 
-# First, add a Furniture class
-class Furniture:
-    def __init__(self, description, liftable=False, hidden_items=None):
-        self.description = description
-        self.liftable = liftable
-        self.hidden_items = hidden_items or []  # Items hidden under this furniture
-        self.is_lifted = False  # Track if furniture has been lifted
-
-@dataclass
-class Position:
-    preposition: str  # on, under, beside, hanging on
-    reference_item: str = None  # what item this is relative to
-
-class Container:
-    def __init__(self, name, description, locked=False, combination=None, key_item=None):
-        self.name = name
-        self.description = description
-        self.items = []
-        self.locked = locked
-        self.combination = combination  # For combination locks
-        self.key_item = key_item       # For key-based locks
-        self.is_open = False
-
-class Room:
-    def __init__(self, name, description):
-        self.name = name
-        self.description = description
-        self.exits = {}
-        self.items = []
-        self.containers = {}  # New dictionary to store containers
-        self.requires = None  # For rooms that require specific items to enter
-
-class Item:
-    def __init__(self, name, description, readable=False, content=None, revealed_clue=None):
-        self.name = name
-        self.description = description
-        self.readable = readable
-        self.content = content
-        self.revealed_clue = revealed_clue
-
-class EnhancedItem(Item):
-    def __init__(self, name, description, readable=False, content=None, revealed_clue=None, position=None, reveals=None, size="medium"):
-        super().__init__(name, description, readable, content, revealed_clue)
-        self.position = position  # Position object
-        self.reveals = reveals or []   # List of items revealed when this item is taken
-        self.size = size         # small, medium, large
-
-class EnhancedRoom(Room):
-    def __init__(self, name, description, items_dict=None, persons={}):
-        super().__init__(name, description)
-        self.revealed_items = set()
-        self.furniture = {}
-        self.persons = persons
-        self.items_dict = items_dict or {}  # Store reference to game's items dictionary
-    
-    def describe_furniture_and_items(self):
-        """Returns a naturally worded description of the room's furniture and visible items"""
-        descriptions = []
-        
-        # First add the base room description
-        # descriptions.append(self.description)
-        
-        # Then describe each piece of furniture and its associated items
-        furniture_items = {}
-        standalone_items = []
-
-        persons = {}
-        standalone_persons = []
-        
-        # # Group items by their furniture reference
-        for item_name in self.items:
-            item = self.items_dict[item_name]
-            if hasattr(item, 'position') and item.position:
-                # Skip hidden items that haven't been revealed
-                if item.position.preposition == "under" and item_name not in self.revealed_items:
-                    continue
-                
-                ref = item.position.reference_item
-                if ref not in furniture_items:
-                    furniture_items[ref] = []
-                furniture_items[ref].append(item_name)
-            else:
-                standalone_items.append(item_name)
-
-        # # Describe each piece of furniture and its items
-        # for furniture_name, furniture_desc in self.furniture.items():
-        #     furniture_text = [furniture_desc]
-        #     if furniture_name in furniture_items:
-        #         items = furniture_items[furniture_name]
-        #         if items:
-        #             preposition = self.items_dict[items[0]].position.preposition
-        #             item_descriptions = [self.items_dict[i].description for i in items]
-        #             furniture_text.append(f"{preposition.capitalize()} it, you see {self._list_to_natural_language(item_descriptions)}.")
-        #     descriptions.append(" ".join(furniture_text))
-
-        # # Add standalone items
-        # if standalone_items:
-        #     descriptions.append("In the room, you also see " + 
-        #                      self._list_to_natural_language([self.items_dict[i].description for i in standalone_items]) + ".")
-
-        # # Add revealed items
-        # revealed = [i for i in self.revealed_items if i in self.items]
-        # if revealed:
-        #     descriptions.append("You can now see " + 
-        #                      self._list_to_natural_language([self.items_dict[i].description for i in revealed]) + ".")
-
-        # return "\n\n".join(descriptions)
-        # Then describe each piece of furniture and its associated items
-        for furniture_name, furniture in self.furniture.items():
-            furniture_text = [furniture.description]
-            if furniture.is_lifted:
-                furniture_text.append("It has been moved from its original position.")
-            if furniture_name in furniture_items:
-                items = furniture_items[furniture_name]
-                if items:
-                    preposition = self.items_dict[items[0]].position.preposition
-                    item_descriptions = [self.items_dict[i].description for i in items]
-                    furniture_text.append(f"{preposition.capitalize()} it, you see {self._list_to_natural_language(item_descriptions)}.")
-            descriptions.append(" ".join(furniture_text))
-
-        #Describe persons
-        for person_name, person in self.persons.items():
-            person_text = []
-            if person.position:
-                ref = person.position.reference_item
-                article = "an" if person_name[0] in list("aeiou") else "a"
-                if ref in self.furniture:
-                    article = "an" if person_name[0] in list("aeiou") else "a"
-                    person_text.append(f"{article} {person_name} is {person.position.preposition} the {self.furniture[ref].lower()}.")
-                elif ref in self.items_dict:
-                    item = self.items_dict[ref]
-                    item_article = "an" if item.description[0] in list("aeiou") else "a"
-                    person_text.append(f"{article} {person_name} is {person.position.preposition} {item_article} {self.items_dict[ref].description}.")
-                else:
-                    person_text.append(f"{article} {person_name} is {person.position.preposition} the {ref}.")
-            person_text.append(person.description)
-            descriptions.append(" ".join(person_text))
-        
-        if standalone_items:
-            descriptions.append("In the room, you also see " + 
-                             self._list_to_natural_language([self.items_dict[i].description for i in standalone_items]) + ".")
-
-        # Add visible items
-        # visible_items = [i for i in self.items if i in self.items_dict]
-        # if visible_items:
-        #     descriptions.append("You can see " + 
-        #                      self._list_to_natural_language([self.items_dict[i].description for i in visible_items]) + ".")
-
-        return "\n\n".join(descriptions)
-
-    def get_descriptive_item_list(self):
-        """Returns a naturally worded description of visible items in the room"""
-        descriptions = []
-        
-        # Group items by their positions
-        hanging_items = []
-        on_items = {}
-        under_items = {}
-        beside_items = {}
-        standalone_items = []
-        
-        for item_name in self.items:
-            item = self.items_dict[item_name]
-            # Skip items that should be hidden
-            if hasattr(item, 'position') and item.position:
-                if item.position.preposition == "under" and item_name not in self.revealed_items:
-                    continue
-                    
-                if item.position.preposition == "hanging on":
-                    hanging_items.append(item_name)
-                elif item.position.preposition == "on":
-                    ref = item.position.reference_item
-                    on_items.setdefault(ref, []).append(item_name)
-                elif item.position.preposition == "under":
-                    ref = item.position.reference_item
-                    under_items.setdefault(ref, []).append(item_name)
-                elif item.position.preposition == "beside":
-                    ref = item.position.reference_item
-                    beside_items.setdefault(ref, []).append(item_name)
-            else:
-                standalone_items.append(item_name)
-
-        # Create natural language descriptions
-        if hanging_items:
-            descriptions.append("Hanging on the walls, you see " + 
-                             self._list_to_natural_language([self.items_dict[i].description for i in hanging_items]))
-        
-        for furniture, items in on_items.items():
-            if furniture in self.furniture:
-                descriptions.append(f"{self.furniture[furniture]} On it, you see " + 
-                                 self._list_to_natural_language([self.items_dict[i].description for i in items]))
-        
-        for furniture, items in beside_items.items():
-            if furniture in self.furniture:
-                descriptions.append(f"Beside {self.furniture[furniture].lower()}, you notice " + 
-                                 self._list_to_natural_language([self.items_dict[i].description for i in items]))
-        
-        revealed = [i for i in self.revealed_items if i in self.items]
-        if revealed:
-            descriptions.append("You can now see " + 
-                             self._list_to_natural_language([self.items_dict[i].description for i in revealed]))
-        
-        if standalone_items:
-            descriptions.append("In the room, you also see " + 
-                             self._list_to_natural_language([self.items_dict[i].description for i in standalone_items]))
-        
-        return "\n".join(descriptions) if descriptions else "There's nothing of particular interest here."
-
-    def _list_to_natural_language(self, items):
-        if not items:
-            return ""
-        if len(items) == 1:
-            return items[0]
-        if len(items) == 2:
-            return f"{items[0]} and {items[1]}"
-        return ", ".join(items[:-1]) + f", and {items[-1]}"
-
-    def describe_containers(self):
-        """Returns a natural description of visible containers"""
-        descriptions = []
-        for name, container in self.containers.items():
-            status = "open" if container.is_open else "closed"
-            desc = f"{container.description} The {name} is {status}."
-            if container.is_open and container.items:
-                item_descs = [self.items_dict[i].description for i in container.items]
-                desc += f" Inside, you can see {self._list_to_natural_language(item_descs)}."
-            descriptions.append(desc)
-        return "\n".join(descriptions) if descriptions else ""
-
-
-from enum import Enum
-
-class InteractionType(Enum):
-    PASSIVE = 1
-    ACTIVE = 2
-
-    @staticmethod
-    def from_string(label):
-        if label.lower() == "passive":
-            return InteractionType.PASSIVE
-        elif label.lower() == "active":
-            return InteractionType.ACTIVE
-        else:
-            raise ValueError("Invalid InteractionType label")
-            
-
-class InteractionAction(Enum):
-    NEXT = 1
-    PREVIOUS = 2
-    END = 3
-
-    @staticmethod
-    def from_string(label):
-        if label.lower() == "next":
-            return InteractionAction.NEXT
-        elif label.lower() == "previous":
-            return InteractionAction.PREVIOUS
-        elif label.lower() == "end":
-            return InteractionAction.END
-        else:
-            raise ValueError("Invalid IteractionAction label")
-
-class InteractionResponse:
-    def __init__(self, response, actions=[InteractionAction.END]):
-        self.response = response
-        self.actions = actions
-
-
-class Interaction:
-    def __init__(self, name, prompt, interaction_type=InteractionType.PASSIVE, responses=[]):
-        self.name = name
-        self.prompt = prompt
-        self.interaction_type = interaction_type
-        self.responses = responses
-
-class Person:
-    def __init__(self, name, description, position=None, dialogues=[]):
-        self.name = name
-        self.description = description
-        self.dialogues = dialogues
-        self.position = position
+from lib.command_parser import CommandParser
+from text_adventure.game_objects import Container, EnhancedItem, EnhancedRoom, Furniture, Interaction, InteractionAction, InteractionResponse, InteractionType, Item, Person, Position, Room
 
 class TextAdventureGame(Game):
     def __init__(self, layout_file="game_layouts.json"):
@@ -320,6 +40,7 @@ class TextAdventureGame(Game):
             "up": ["up", "u", "upward", "upstairs", "climb up"],
             "down": ["down", "d", "downward", "downstairs", "climb down"]
         }
+        self.parser = CommandParser()
         self.setup_game()
 
     def load_layouts(self):
@@ -415,14 +136,7 @@ class TextAdventureGame(Game):
 
     def _identify_command(self, user_input):
         """Identifies the core command from user input"""
-        words = user_input.lower().split()
-        
-        for word in words:
-            for command, patterns in self._command_patterns.items():
-                if word in patterns:
-                    return command, words
-        
-        return None, words
+        return self.parser.parse_command(user_input)
 
     def _identify_direction(self, words):
         """Identifies direction from a list of words"""
@@ -432,9 +146,9 @@ class TextAdventureGame(Game):
                     return direction
         return None
 
-    def _identify_item(self, words, available_items):
+    def _identify_item(self, command, available_items):
         """Identifies an item from user input"""
-        user_phrase = " ".join(words)
+        user_phrase = command['direct_object']
         # Check for full item names first
         for item in available_items:
             if item.lower() in user_phrase:
@@ -443,7 +157,7 @@ class TextAdventureGame(Game):
         # Check for partial matches
         for item in available_items:
             words_in_item = item.lower().split()
-            for word in words:
+            for word in user_phrase.split():
                 if word in words_in_item:
                     return item
         
@@ -452,83 +166,85 @@ class TextAdventureGame(Game):
     def get_command(self):
         """Gets and processes user command with natural language understanding"""
         raw_input = input("> ").lower()
-        command, words = self._identify_command(raw_input)
+        command = self._identify_command(raw_input)
         
-        if not command:
+        if not command['action']:
             print("\n❓ I'm not sure what you want to do.")
             print("   Try using simple commands like 'go', 'take', or 'read'")
             print("   Type 'help' for a list of commands.")
             return
             
-        if command == "quit":
+        if command["action"] == "quit":
             self.is_running = False
-        elif command == "inventory":
+        elif command["action"] == "inventory":
             self.show_inventory()
-        elif command == "objective":
+        elif command["action"] == "objective":
             self.show_objective()
-        elif command == "go":
-            direction = self._identify_direction(words)
+        elif command["action"] == "go":
+            direction = self._identify_direction(command["direction"])
             if direction:
                 self.move_player(direction)
             else:
                 print("\n❓ Which direction do you want to go?")
-        elif command == "take":
+        elif command["action"] == "take":
             #combine the items in the room and the items in the containers
-            items = self._current_room.items + [item for container in self._current_room.containers.values() for item in container.items if container.is_open]
-            item = self._identify_item(words, items)
+            items = [item['name'] for item in self._current_room.items] + [item for container in self._current_room.containers.values() for item in container.items if container.is_open]
+            item = self._identify_item(command, items)
             if item:
                 self.take_item(item)
             else:
                 print("\n❌ What do you want to take?")
-        elif command == "drop":
-            item = self._identify_item(words, self._inventory)
+        elif command["action"] == "drop":
+            item = self._identify_item(command, self._inventory)
             if item:
                 self.drop_item(item)
             else:
                 print("\n❌ What do you want to drop?")
-        elif command in ["read", "examine"]:
-            item_name = self._identify_item(words, self._inventory + self._current_room.items)
+        elif command["action"] in ["read", "examine"]:
+            item_name = self._identify_item(command, self._inventory + self._current_room.items)
             if item_name:
                 self.read_item(item_name)
             else:
                 print("\n❌ What do you want to read?")
-        elif command == "open":
-            container_name = self._identify_container(words)
+        elif command["action"] == "open":
+            container_name = self._identify_container(command)
             if container_name:
                 self.open_container(container_name)
             else:
                 print("\n❓ What do you want to open?")
         
-        elif command == "close":
-            container_name = self._identify_container(words)
+        elif command["action"] == "close":
+            container_name = self._identify_container(command)
             if container_name:
                 self.close_container(container_name)
             else:
                 print("\n❓ What do you want to close?")
         
-        elif command == "put":
-            item = self._identify_item(words, self._inventory)
-            container_name = self._identify_container(words)
+        elif command["action"] == "put":
+            item = self._identify_item(command, self._inventory)
+            container_name = self._identify_container(command)
             if item and container_name:
                 self.put_item_in_container(item, container_name)
             else:
                 print("\n❓ Please specify both an item and a container.")
         
-        elif command == "combine":
-            container_name = self._identify_container(words)
+        elif command["action"] == "combine":
+            container_name = self._identify_container(command)
             # Look for numbers in the input
-            numbers = [word for word in words if word.isdigit()]
+            numbers = [word for word in command["direct_object"].split() if word.isdigit()]
+            if len(numbers) == 0:
+                numbers = [word for word in command["indirect_object"].split() if word.isdigit()]
             if container_name and numbers:
                 self.try_combination(container_name, "".join(numbers))
             else:
                 print("\n❓ Please specify a container and combination.")
 
-        elif command == "help":
+        elif command["action"] == "help":
             self.show_help()
     
-    def _identify_container(self, words):
+    def _identify_container(self, command):
         """Identifies a container from user input"""
-        user_phrase = " ".join(words)
+        user_phrase = command['direct_object']
         available_containers = self._current_room.containers.keys()
         
         for container in available_containers:
@@ -727,7 +443,7 @@ class TextAdventureGame(Game):
             print("\n💡 You see:")
             if self._current_room.items:
                 for item in self._current_room.items:
-                    print(f"   📦 {item}")
+                    print(f"   📦 {item['name']}")
             if self._current_room.containers:
                 for name, container in self._current_room.containers.items():
                     status = "🔓 (open)" if container.is_open else "🔒 (closed)"
@@ -746,8 +462,9 @@ class TextAdventureGame(Game):
         print("═" * 50)
     
     def take_item(self, item):
-        if item in self._current_room.items:
-            self._current_room.items.remove(item)
+        found_item = self._current_room.find_item(item)
+        if found_item:
+            self._current_room.remove_item(item)
             self._inventory.append(item)
             print(f"\n✅ You take the {item}.")
             if self.check_win_condition():  # Check win condition after taking item
@@ -772,7 +489,7 @@ class TextAdventureGame(Game):
     def drop_item(self, item):
         if item in self._inventory:
             self._inventory.remove(item)
-            self._current_room.items.append(item)
+            self._current_room.items.append({'name': item})
             print(f"\n✅ You drop the {item}.")
             if self.check_win_condition():  # Check win condition after dropping item (in case it affects the win state)
                 self._completed = True
@@ -835,6 +552,7 @@ class TextAdventureGame(Game):
 
 class EnhancedTextAdventureGame(TextAdventureGame):
     def __init__(self, layout_file="game_layouts.json"):
+        super().__init__(layout_file)
         self._inventory = []
         self._rooms = {}
         self._current_room = None
@@ -890,35 +608,47 @@ class EnhancedTextAdventureGame(TextAdventureGame):
 
         # Create items
         for item_name, item_data in layout['items'].items():
-            position_data = item_data.get('position', {})
+            # position_data = item_data.get('position', {})
             self._items[item_name] = EnhancedItem(
                 item_name,
                 item_data['description'],
                 item_data.get('readable', False),
                 item_data.get('content'),
                 item_data.get('revealed_clue'),
-                Position(
-                    position_data.get('preposition', None),
-                    position_data.get('reference', None)
-                ) if position_data else None,
+                None,
                 item_data.get('reveals', []),
                 item_data.get('size', 'medium')
             )
 
         # Create rooms with furniture
         for room_id, room_data in layout['rooms'].items():
-            room = EnhancedRoom(room_data['name'], room_data['description'], self._items)
+
+            room_items_dict = {}
+            for item in room_data['items']:
+                item_name = item['name']
+                room_obj = self._items[item_name]
+                position_data = item.get('position', {})
+                room_obj.position = Position(
+                    position_data.get('preposition', None),
+                    position_data.get('reference', None)
+                )
+                room_items_dict[item_name] = room_obj
+
+            room = EnhancedRoom(room_data['name'], room_data['description'], room_items_dict)
             room.exits = room_data['exits']
             room.items = room_data['items']
+
+            
             
             # Create Furniture objects
             for furniture_name, furniture_data in room_data.get('furniture', {}).items():
                 if isinstance(furniture_data, str):
                     # Handle old format where furniture was just a description
-                    furniture = Furniture(furniture_data, liftable=False)
+                    furniture = Furniture('', furniture_data, liftable=False)
                 else:
                     # Handle new format with full furniture data
                     furniture = Furniture(
+                        furniture_data['name'] if 'name' in furniture_data else '',
                         furniture_data['description'],
                         liftable=furniture_data.get('liftable', False),
                         hidden_items=furniture_data.get('hidden_items', [])
@@ -1005,14 +735,15 @@ class EnhancedTextAdventureGame(TextAdventureGame):
 
     def take_item(self, item):
         """Enhanced take item that reveals hidden items"""
-        if item in self._current_room.items:
+        found_item = self._current_room.find_item(item)
+        if found_item:
             item_obj = self._items[item]
-            if hasattr(item_obj, 'position') and item_obj.position:
-                if item_obj.position.preposition == "under" and item not in self._current_room.revealed_items:
+            if found_item['position']:
+                if found_item['position']['preposition'] == "under" and item not in self._current_room.revealed_items:
                     print(f"\n❌ You don't see the {item} anywhere obvious.")
                     return
             
-            self._current_room.items.remove(item)
+            self._current_room.remove_item(item)
             self._inventory.append(item)
             print(f"\n✅ You take the {item}.")
             
@@ -1062,7 +793,7 @@ class EnhancedTextAdventureGame(TextAdventureGame):
         
         if furniture.hidden_items:
             # Add hidden items to the room
-            self._current_room.items.extend(furniture.hidden_items)
+            self._current_room.add_items(furniture.hidden_items)
             # Add to revealed items
             self._current_room.revealed_items.update(furniture.hidden_items)
             items_desc = ", ".join(furniture.hidden_items)
@@ -1107,90 +838,84 @@ class EnhancedTextAdventureGame(TextAdventureGame):
     def get_command(self):
         """Gets and processes user command with natural language understanding"""
         raw_input = input("> ").lower()
-        command, words = self._identify_command(raw_input)
+        command = self._identify_command(raw_input)
         
-        if not command:
+        if not command['action']:
             print("\n❓ I'm not sure what you want to do.")
             print("   Try using simple commands like 'go', 'take', or 'read'")
             print("   Type 'help' for a list of commands.")
             return
             
-        if command == "quit":
+        if command["action"] == "quit":
             self.is_running = False
-        elif command == "inventory":
+        elif command["action"] == "inventory":
             self.show_inventory()
-        elif command == "objective":
+        elif command["action"] == "objective":
             self.show_objective()
-        elif command == "go":
-            direction = self._identify_direction(words)
+        elif command["action"] == "go":
+            direction = self._identify_direction(command["direction"])
             if direction:
                 self.move_player(direction)
             else:
                 print("\n❓ Which direction do you want to go?")
-        elif command == "take":
+        elif command["action"] == "take":
             #combine the items in the room and the items in the containers
-            items = self._current_room.items + [item for container in self._current_room.containers.values() for item in container.items if container.is_open]
-            print(items)
-            item = self._identify_item(words, items)
+            items = [item['name'] for item in self._current_room.items] + [item for container in self._current_room.containers.values() for item in container.items if container.is_open]
+            item = self._identify_item(command, items)
             if item:
                 self.take_item(item)
             else:
                 print("\n❌ What do you want to take?")
-        elif command == "drop":
-            item = self._identify_item(words, self._inventory)
-            if item:
-                self.drop_item(item)
-            else:
-                print("\n❌ What do you want to drop?")
-        elif command in ["read", "examine"]:
-            item_name = self._identify_item(words, self._inventory + self._current_room.items)
+        elif command["action"] == "drop":
+                item = self._identify_item(command, self._inventory)
+                if item:
+                    self.drop_item(item)
+                else:
+                    print("\n❌ What do you want to drop?")
+        elif command["action"] in ["read", "examine"]:
+            item_name = self._identify_item(command, self._inventory + [item['name'] for item in self._current_room.items])
             if item_name:
                 self.read_item(item_name)
             else:
                 print("\n❌ What do you want to read?")
-        elif command == "open":
-            container_name = self._identify_container(words)
+        elif command["action"] == "open":
+            container_name = self._identify_container(command)
             if container_name:
                 self.open_container(container_name)
             else:
                 print("\n❓ What do you want to open?")
         
-        elif command == "close":
-            container_name = self._identify_container(words)
+        elif command["action"] == "close":
+            container_name = self._identify_container(command)
             if container_name:
                 self.close_container(container_name)
             else:
                 print("\n❓ What do you want to close?")
         
-        elif command == "put":
-            item = self._identify_item(words, self._inventory)
-            container_name = self._identify_container(words)
+        elif command["action"] == "put":
+            item = self._identify_item(command, self._inventory)
+            container_name = self._identify_container(command)
             if item and container_name:
                 self.put_item_in_container(item, container_name)
             else:
                 print("\n❓ Please specify both an item and a container.")
         
-        elif command == "combine":
-            container_name = self._identify_container(words)
+        elif command["action"] == "combine":
+            container_name = self._identify_container(command)
             # Look for numbers in the input
-            numbers = [word for word in words if word.isdigit()]
+            numbers = [word for word in command["direct_object"].split() if word.isdigit()]
+            if len(numbers) == 0:
+                numbers = [word for word in command["indirect_object"].split() if word.isdigit()]
             if container_name and numbers:
                 self.try_combination(container_name, "".join(numbers))
             else:
                 print("\n❓ Please specify a container and combination.")
-
-        elif command == "lift":
-            # First check if it's a piece of furniture
-            furniture_name = next((f for f in self._current_room.furniture.keys() 
-                                if f.lower() in " ".join(words).lower()), None)
-            if furniture_name:
-                self.lift_furniture(furniture_name)
-        elif command == "help":
+        elif command["action"] == "help":
             self.show_help()
-        elif command == "lift":
+        elif command["action"] == "lift":
             # First check if it's a piece of furniture
             furniture_name = next((f for f in self._current_room.furniture.keys() 
-                                if f.lower() in " ".join(words).lower()), None)
+                                if f.lower() in " ".join(command["direct_object"]).lower()), None)
             if furniture_name:
                 self.lift_furniture(furniture_name)
             else:
